Add without sub directive to fastcgi
To allow easy handling of sub urls, this change adds the `without`
directive to fastcgi. When set and a path has the without value as
prefix, this prefix is stripped from the path before creating the
fastcgi environment.

Example:

fastcgi2 /subpath/ 127.0.0.1:9000 php {
  without /subpath/
  root /usr/share/my-php-app
}

This is useful, if a fastcgi app is to be served in a subpath.