Add support for extra configuration folder
Previously extra parameter had to be a single file. Since this limits
the flexibility, this change also support the value to be a
directory. If it is a directory, all file inside it with the `.cfg`
suffix will be loaded in alphabetical order and added to kweb
configuration. If extra value is a file, behavior remains unchanged.