Fix alias directive redirect targets
When using the alias directive with a prefix and the static request
results in an redirect (like to index.html) or to a folder but without
the trailing slash, the resulting redirect URL is wrong. This change
retains the prefix path to make it available properly when generating
the redirect location.