Fix request_log folder creation check to avoid changing owner
It was found that when using an existing folder with the
request_log_file configuration, that folders owner is changed to the
kwebd runtime user if the request log file does not exist. This change
modifies this behavior so the owner of the folder is only changed if it
actually also created by kwebd.